dj_map/requirements.txt
celery==5.4.0
Django==5.0.4
djangorestframework==3.15.1
pillow==10.3.0
python-dateutil==2.9.0.post0
redis==4.5.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
